refactor(agents): own http runtime scaffold
