+++ services/agents/package.json
@@ -7,6 +7,7 @@
     "./server/controller-startup-retry": "./src/server/controller-startup-retry.ts",
     "./server/env-compat": "./src/server/env-compat.ts",
     "./server/http": "./src/server/http.ts",
+    "./server/http-runtime": "./src/server/http-runtime.ts",
     "./server/kube-gateway": "./src/server/kube-gateway.ts",
     "./server/kube-types": "./src/server/kube-types.ts",
     "./server/kubernetes-bun-transport": "./src/server/kubernetes-bun-transport.ts",
@@ -32,7 +33,9 @@
     "@kubernetes/client-node": "^1.4.0",
     "@proompteng/codex": "workspace:*",
     "@proompteng/temporal-bun-sdk": "workspace:*",
+    "crossws": "^0.4.5",
     "effect": "3.21.2",
+    "h3": "2.0.1-rc.22",
     "xstate": "5.31.0"
   },
   "devDependencies": {
